Update import paths for user expense and user fetching API schemas, and remove unused globals.css file to streamline the codebase. This change enhances maintainability by consolidating schema imports and eliminating obsolete styles.
