+++ src/api/user/fetch-user-expenses.ts
@@ -1,4 +1,4 @@
-import { userExpensesResponseSchema } from "@/app/(home)/data-table/schema";
+import { userExpensesResponseSchema } from "@/app/(home)/data-table/schema/schema";
 
 const API_BASE_URL = "/api";
