+++ src/api/user/fetch-users.ts
@@ -1,4 +1,4 @@
-import { usersResponseSchema } from "@/app/(home)/data-table/schema";
+import { usersResponseSchema } from "@/app/(home)/data-table/schema/schema";
 
 const API_BASE_URL = "/api";
