+++ src/app/(home)/data-table/schema/expense-schema.ts
@@ -1,20 +1,8 @@
 import { z } from "zod";
 
-// Schema for users with expenses from the API
-export const userSchema = z.object({
-  id: z.number(),
-  name: z.string(),
-  email: z.string(),
-  phone: z.string(),
-  age: z.number(),
-  created_at: z.string(),
-  expense_count: z.number(),
-  total_expenses: z.string(),
-});
+// ** Import Schema
+import { userSchema } from "./user-schema";
 
-export type User = z.infer<typeof userSchema>;
-
-// Schema for expense data from API
 export const expenseSchema = z.object({
   id: z.number(),
   user_id: z.number(),
@@ -26,18 +14,6 @@ export const expenseSchema = z.object({
 });
 
 export type Expense = z.infer<typeof expenseSchema>;
-
-// API response schemas
-export const usersResponseSchema = z.object({
-  success: z.boolean(),
-  data: z.array(userSchema),
-  pagination: z.object({
-    page: z.number(),
-    limit: z.number(),
-    total_pages: z.number(),
-    total_items: z.number(),
-  }),
-});
 
 export const userExpensesResponseSchema = z.object({
   success: z.boolean(),
@@ -57,4 +33,4 @@ export const userExpensesResponseSchema = z.object({
     total_pages: z.number(),
     total_items: z.number(),
   }),
-});
+});
