+++ src/app/(home)/data-table/schema/index.ts
@@ -1,0 +1,2 @@
+export * from "./user-schema";
+export * from "./expense-schema";
